refactor: model 계층 분리
