+++ src/index.js
@@ -1,62 +1,14 @@
   import NumberUtils from "./utils/NumberUtils.js";
   import BaseballGameView from "./view/BaseBallGameView.js";
+  import BaseballGame from "./model/BaseBallGame.js";
 
-  export default class BaseballGame {
+  export default class BaseballGameController {
 
     constructor() {
-      this.computerNumbers = [];
       this.NumberUtils = new NumberUtils();
       this.BaseballGameView = new BaseballGameView();
+      this.BaseballGame = new BaseballGame();
       this.gameStart();
-    }
-
-    play(computerNumbers, userNumbers) {
-      const strikeAndBallCounts = this.compareNumbers(userNumbers, computerNumbers);
-      return this.extractResult(strikeAndBallCounts);
-    }
-
-    extractResult({ ball, strike }) {
-      if(ball === 0 && strike === 0)
-        return "낫싱";
-      if(ball !== 0 && strike === 0)
-        return `${ball}볼`;
-      if(ball === 0 && strike !== 0)
-        return `${strike}스트라이크`;
-      return `${ball}볼 ${strike}스트라이크`;
-    }
-
-    compareNumbers(userNumbers, computerNumbers) {
-      let strikeAndBallCounts = {
-        strike : 0,
-        ball : 0
-      };
-
-      for(let i = 0; i < 3; i++) {
-        if(userNumbers[i] === computerNumbers[i]) 
-          strikeAndBallCounts.strike += 1;
-        else if(computerNumbers.includes(userNumbers[i]))
-          strikeAndBallCounts.ball += 1;
-      }
-
-      return strikeAndBallCounts;
-    }
-
-    makeComputerNumbers() {
-      let computerNumbers = [];
-
-      while(!this.NumberUtils.validateDifferentNumber(computerNumbers)) {
-        computerNumbers = [
-          this.makeRandomNumber(),
-          this.makeRandomNumber(),
-          this.makeRandomNumber()
-        ];
-      }
-
-      return computerNumbers;
-    }
-
-    makeRandomNumber() {
-      return MissionUtils.Random.pickNumberInRange(1, 9);
     }
 
     endGame(result) {
@@ -72,7 +24,7 @@
     }
 
     restart() {
-      this.computerNumbers = this.makeComputerNumbers();
+      this.BaseballGame.scomputerNumbers = this.BaseballGame.makeComputerNumbers();
       this.BaseballGameView.resetInput();
       this.BaseballGameView.resetResult();
       this.BaseballGameView.toggleButtons(true);
@@ -82,12 +34,12 @@
       const input = document.querySelector("#user-input");
       const submitButton = document.querySelector("#submit");
 
-      this.computerNumbers = this.makeComputerNumbers();
+      this.BaseballGame.computerNumbers = this.BaseballGame.makeComputerNumbers();
 
       submitButton.addEventListener('click', (event) => {
         event.preventDefault();
         
-        console.log(this.computerNumbers);
+        console.log(this.BaseballGame.computerNumbers);
 
         const userInput = input.value;
         const userNumbers = userInput.split('').map(Number);
@@ -98,7 +50,7 @@
           return;
         }
 
-        const result = this.play(userNumbers, this.computerNumbers);
+        const result = this.BaseballGame.play(userNumbers, this.BaseballGame.computerNumbers);
         this.BaseballGameView.printResult(result);
 
         if(this.NumberUtils.isPlayerWinner(result)){
@@ -108,4 +60,4 @@
     }
   }
 
-new BaseballGame();
+new BaseballGameController();
